+++ src/bpms/bpmn/elements/BusinessRuleTask.ts
@@ -1,5 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-// tslint:disable-next-line: no-var-requires
-const { ServiceTask: BusinessRuleTask } = require('bpmn-elements');
-
-export default BusinessRuleTask;