+++ src/bpms/bpmn/extensions/UserTaskExtension.ts
@@ -8,15 +8,35 @@ export const UserTaskExtension = (processInstance: BpmnProcessInstance) => (acti
         activity.on('wait', async (api: BpmnProcessActivity) => {
             const bpms = processInstance.BpmnEngine.BpmsEngine;
             if (bpms) {
-                const t = await bpms.TaskService.create({
-                    name: api.name,
-                    activityId: api.id,
-                    activityType: api.type,
-                    processDefinitionId: processInstance.DefinitionId,
-                    processInstanceName: processInstance.Name,
-                    processInstanceId: processInstance.Id,
-                    tenantId: bpms.Name,
-                });
+                const { fields, content, properties, ...variables } = api.environment.variables;
+                const initiator = variables.initiator;
+                const assignees: { userId: string; username: string }[] = [];
+                const potentialOwner = api.content.potentialOwner;
+                const performerUsers = api.content.humanPerformer && (await api.content.humanPerformer(api));
+                if (Array.isArray(performerUsers)) {
+                    for (const user of performerUsers) {
+                        assignees.push(user);
+                    }
+                } else {
+                    assignees.push(performerUsers || initiator);
+                }
+                for (const assignee of assignees) {
+                    const t = await bpms.TaskService.create({
+                        name: api.name,
+                        activityId: api.id,
+                        activityType: api.type,
+                        processDefinitionId: processInstance.DefinitionId,
+                        processDefinitionName: processInstance.DefinitionName,
+                        processDefinitionVersion: processInstance.DefinitionVersion,
+                        processInstanceName: processInstance.Name,
+                        processInstanceId: processInstance.Id,
+                        tenantId: bpms.Name,
+                        priority: 'normal',
+                        descriptions: '',
+                        variables: variables,
+                        assignee: assignee,
+                    });
+                }
             }
         });
     }